rename group.py | solve circular import
