+++ docs/notebooks/4d_Predict_emergency_demand_with_special_categories.md
@@ -16,6 +16,9 @@ Most of the code below is the same as in the previous notebook. I limit the narr
 %autoreload 2
 ```
 
+    The autoreload extension is already loaded. To reload it, use:
+      %reload_ext autoreload
+
 ```python
 from patientflow.load import set_project_root
 project_root = set_project_root()
@@ -618,26 +621,28 @@ apply_special_category_filtering=True,
 admit_col=&#x27;is_admitted&#x27;
 )</pre></div> </div></div></div></div>
 
-Under the hood, the `SequenceToOutcomePredictor` will call a `create_special_category_objects()` function that returns rules for how to handle each subgroup. The implementation here is primarily designed to handle paediatric patients (under 18) as a special category. A `SpecialCategoryParams` class generates a dictionary mapping specialties to flags (1.0 for paediatric, 0.0 for others) and functions to identify paediatric patients based on age data. It provides methods to handle both age formats (age_on_arrival or age_group).
+Under the hood, the `SequenceToOutcomePredictor` will call a `create_special_category_objects()` function that returns rules for how to handle each subgroup. The implementation here is primarily designed to handle pediatric patients (under 18) as a special category. A `SpecialCategoryParams` class generates a dictionary mapping specialties to flags (1.0 for pediatric, 0.0 for others) and functions to identify pediatric patients based on age data. It provides methods to handle both age formats (age_on_arrival or age_group).
 
 The `SequenceToOutcomePredictor` applies these rules during both training and prediction, ensuring consistent handling of special categories across the entire prediction pipeline
 
 The `SpecialCategoryParams` class is designed to be picklable, which is necessary for saving the specialty predictor model to disk.
 
-The output from `create_special_category_objects` is shown below. Note that the output is specific to the UCLH implementation. See below for notes about how to change this for your implementation.
+The output from `create_special_category_objects` is shown below. Note that the output is specific to the UCLH implementation, and that its use has been deprecated in favour of a more generalisable approach.
+
+See below for notes about how to change this for your implementation.
 
 ```python
 from patientflow.prepare import create_special_category_objects
 create_special_category_objects(train_visits_df.columns)
 ```
 
-    {'special_category_func': <bound method SpecialCategoryParams.special_category_func of <patientflow.prepare.SpecialCategoryParams object at 0x12fcba2c0>>,
+    {'special_category_func': <function patientflow.group.create_special_category_objects.<locals>.is_pediatric(row)>,
      'special_category_dict': {'medical': 0.0,
       'surgical': 0.0,
       'haem/onc': 0.0,
       'paediatric': 1.0},
-     'special_func_map': {'paediatric': <bound method SpecialCategoryParams.special_category_func of <patientflow.prepare.SpecialCategoryParams object at 0x12fcba2c0>>,
-      'default': <bound method SpecialCategoryParams.opposite_special_category_func of <patientflow.prepare.SpecialCategoryParams object at 0x12fcba2c0>>}}
+     'special_func_map': {'paediatric': <function patientflow.group.create_special_category_objects.<locals>.is_pediatric(row)>,
+      'default': <function patientflow.group.create_special_category_objects.<locals>.is_not_pediatric(row)>}}
 
 ## Train models for yet-to-arrive patients
 
@@ -645,7 +650,7 @@ Predictions for patients who are yet-to-arrive models are based on arrival rates
 
 The `create_yta_filters()` function generates a dictionary of filters for the `ParametricIncomingAdmissionPredictor` to enable separate prediction models for each specialty. It uses the same special category configuration (as defined in `create_special_category_objects`) to create two types of filters:
 
-- For paediatric patients: {"is_child": True}
+- For pediatric patients: {"is_child": True}
 - For other specialties: {"specialty": specialty_name, "is_child": False}
 
 This allows the predictor to
@@ -679,7 +684,7 @@ yta_model_by_spec =yta_model_by_spec.fit(train_inpatient_arrivals_df,
 
 ### Saving of special category information
 
-The `ParametricIncomingAdmissionPredictor` class uses the special category objects during initialisation to create static filters that map specialties to their configurations (e.g., {'is_child': True} for paediatric cases), but does not need them in the predict method. The filters are saved with the instance.
+The `ParametricIncomingAdmissionPredictor` class uses the special category objects during initialisation to create static filters that map specialties to their configurations (e.g., {'is_child': True} for pediatric cases), but does not need them in the predict method. The filters are saved with the instance.
 
 ```python
 yta_model_by_spec.filters
@@ -696,17 +701,22 @@ In contrast, the `SequenceToOutcomePredictor` save the special parameters as a f
 spec_model.special_params
 ```
 
-    {'special_category_func': <bound method SpecialCategoryParams.special_category_func of <patientflow.prepare.SpecialCategoryParams object at 0x12fcba3f0>>,
+    {'special_category_func': <function patientflow.group.create_subgroup_functions.<locals>.is_pediatric(row)>,
      'special_category_dict': {'medical': 0.0,
       'surgical': 0.0,
       'haem/onc': 0.0,
       'paediatric': 1.0},
-     'special_func_map': {'paediatric': <bound method SpecialCategoryParams.special_category_func of <patientflow.prepare.SpecialCategoryParams object at 0x12fcba3f0>>,
-      'default': <bound method SpecialCategoryParams.opposite_special_category_func of <patientflow.prepare.SpecialCategoryParams object at 0x12fcba3f0>>}}
+     'special_func_map': {'paediatric': <function patientflow.group.create_subgroup_functions.<locals>.is_pediatric(row)>,
+      'default': <function patientflow.group.create_subgroup_system.<locals>.<lambda>(row)>,
+      'pediatric': <function patientflow.group.create_subgroup_functions.<locals>.is_pediatric(row)>,
+      'adult_male_young': <function patientflow.group.create_subgroup_functions.<locals>.is_adult_male_young(row)>,
+      'adult_female_young': <function patientflow.group.create_subgroup_functions.<locals>.is_adult_female_young(row)>,
+      'adult_male_senior': <function patientflow.group.create_subgroup_functions.<locals>.is_adult_male_senior(row)>,
+      'adult_female_senior': <function patientflow.group.create_subgroup_functions.<locals>.is_adult_female_senior(row)>}}
 
 ## Changes required in your implementation
 
-Listed below are the functions that relate to this special handling.
+Listed below are the functions that relate to this special handling. However, note that the way this is handled has been deprecated, in favour of a more generalisable approach.
 
 1. **SpecialCategoryParams Class** (`src/patientflow/prepare.py`):
 
@@ -1027,3 +1037,5 @@ for specialty in ['medical', 'surgical', 'haem/onc', 'paediatric']:
 ## Summary
 
 In this notebook I have shown how to specify that certain groups are handled differently. In the UCLH case, we assume that all patients under 18 will be admitted to a paediatric specialty. I have demonstrated how you can use the functions in patientflow to handle such special cases.
+
+Note: the approach illustrated here had been deprecated. The code will still work, but this notebook will be replaced in due course.
